Decoupled adaptors from main lib and extracted them into a library extension
