--- src/Liip/Drupal/Modules/Registry/Adaptor/AdaptorException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Liip\Drupal\Modules\Registry\Adaptor;
-
-
-class AdaptorException extends \Exception
-{
-}
